tests/mma8x5x: adapted test for driver changes
