--- tests/driver_mma8652/Makefile
@@ -1,25 +1,0 @@
-APPLICATION = driver_mma8652
-include ../Makefile.tests_common
-
-FEATURES_REQUIRED = periph_i2c
-
-USEMODULE += mma8652
-USEMODULE += xtimer
-
-# set default device parameters in case they are undefined
-TEST_MMA8652_I2C           ?= I2C_DEV\(0\)
-TEST_MMA8652_ADDR          ?= 0x1D
-TEST_MMA8652_USER_OFFSET_X ?= 0
-TEST_MMA8652_USER_OFFSET_Y ?= 0
-TEST_MMA8652_USER_OFFSET_Z ?= 0
-TEST_MMA8x5x_TYPE          ?= MMA8x5x_TYPE_MMA8652
-
-# export parameters
-CFLAGS += -DTEST_MMA8652_I2C=$(TEST_MMA8652_I2C)
-CFLAGS += -DTEST_MMA8652_ADDR=$(TEST_MMA8652_ADDR)
-CFLAGS += -DTEST_MMA8652_USER_OFFSET_X=$(TEST_MMA8652_USER_OFFSET_X)
-CFLAGS += -DTEST_MMA8652_USER_OFFSET_Y=$(TEST_MMA8652_USER_OFFSET_Y)
-CFLAGS += -DTEST_MMA8652_USER_OFFSET_Z=$(TEST_MMA8652_USER_OFFSET_Z)
-CFLAGS += -DTEST_MMA8x5x_TYPE=$(TEST_MMA8x5x_TYPE)
-
-include $(RIOTBASE)/Makefile.include
