+++ tests/driver_mma8x5x/Makefile
@@ -1,0 +1,9 @@
+APPLICATION = driver_mma8x5x
+include ../Makefile.tests_common
+
+FEATURES_REQUIRED = periph_i2c
+
+USEMODULE += mma8x5x
+USEMODULE += xtimer
+
+include $(RIOTBASE)/Makefile.include
